+++ internal/models/bar.go
@@ -10,5 +10,5 @@ type Bar struct {
 	High     float64
 	Low      float64
 	Close    float64
-	Volume   float64
+	Volume   int64
 }
